pyproject.toml
[project]
name = "langgraph_eng_package"
version = "0.0.1"
dependencies = [
  "langgraph",
"langchain_anthropic",
"langchain_core",
"langchain_openai",
"langchain-google-genai",
"python-dotenv"
]

[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"